requirements.txt
flask==3.1.1
gunicorn==21.2.0
python-dotenv==1.1.0
langchain==0.3.26
langchain-pinecone==0.2.8
langchain-groq==0.2.2
langchain-community==0.3.26
sentence-transformers==4.1.0
pypdf==5.6.1
pinecone-client==5.0.1
requests==2.31.0
torch==2.0.1
transformers==4.36.0